Requerimientos cumplidos Algún día se limpiara el codigo -.-

+++ app/src/main/java/com/desafiolatam/restapi/consumo_API/pojo/Productos.kt
@@ -3,7 +3,7 @@ package com.desafiolatam.restapi.consumo_API.pojo
 data class Producto(
     val id: Int,
     val title: String,
-    val price: Int,
+    val price: Long,
     val description: String,
     val images: List<String>,
     val creationAt: String,
